Camera now moves to Interactable
When pressing "E" over an interactable the whole player moves and the camera (via the player controller) is forced to rotate to look at the interactable.

It sort of works but is really clunky. Probably worth trying to the multi-cam method (with SetViewTargetWithBlend()) but just saving here in case it all goes wrong...