Blur With Widget Not Working
Tried blurring the screen with a widget component but it doesn't seem to be activating for some reason.

It probably wouldn't be any use anyway as it would blur the object in the inventory.

Definitely need to be able to do this with a camera instead (or not at all)