Code Sounds Created
The code pad now makes sounds when different combinations are pressed.

Need to figure out why the sound is slow to play the first time round (assume this is because audio component isn't playing. Maybe need to turn them all on when player interacts with component, and off when player leaves/solves code?)