Bug Fixes
Have made a small edit to the if statement that dictates whether button is rising or falling. Hasn't completely eliminated the problem, but it does seem to have helped.

Have set the Audio Components on the Code Machines to fade to 0 instantly. This makes sure the sound is "prewarmed" so it comes in straight away when the key is pressed. This might be computationally expensive with several code machines in the scene but will cross that bridge when we come to it as trying to "activate" them when player interacts with object instead hasn't worked.