Code Symbols Started
Have been playing around with TextRenderComponents to make them appear as keys are triggered.

This is working but have not been able to find a way to make them fade in, might need to create my own material from scratch in C++ so I can have a parameter that controls opacity.

Alternatively, might end up creating static meshes that I can do something with, would be just as easy to make those in Blender as it would to create a font, and I might have more control over fading etc.